compartment

+++ src/global_types.rs
@@ -1,3 +1,3 @@
 pub type PartialPressure = f32;
 pub type Depth = f32;
-pub type Minutes = f64;
+pub type Minutes = f32;
